requirements.txt
openai
python-dotenv
music21
midi2audio
